Address title review comments
