+++ apps/desktop/src/lib/titleGeneration.ts
@@ -2,6 +2,7 @@ import { parseMarkdownFrontMatter } from "@hubble.md/editor";
 
 export const MAX_GENERATED_TITLE_LENGTH = 40;
 
+/** Builds a file-safe stem from the first line with visible text. */
 export function generatedTitleStem(markdown: string): string | null {
 	const body = parseMarkdownFrontMatter(markdown).body;
 	for (const line of body.split("\n")) {
@@ -9,8 +10,11 @@ export function generatedTitleStem(markdown: string): string | null {
 		if (!text) continue;
 		const slug = text
 			.toLocaleLowerCase()
+			// Join contractions instead of treating apostrophes as word breaks.
 			.replace(/['’`]/g, "")
+			// Keep Unicode letters and numbers; turn each other run into one dash.
 			.replace(/[^\p{L}\p{N}]+/gu, "-")
+			// A file stem cannot start or end with a generated separator.
 			.replace(/^-+|-+$/g, "");
 		const capped = Array.from(slug)
 			.slice(0, MAX_GENERATED_TITLE_LENGTH)
@@ -21,9 +25,11 @@ export function generatedTitleStem(markdown: string): string | null {
 	return null;
 }
 
+/** Strips Markdown syntax that does not read as title text. */
 function meaningfulText(line: string): string {
 	const trimmed = line.trim();
 	if (
+		// Horizontal rules, fenced-code markers, and link definitions have no title.
 		/^(?:-{3,}|_{3,}|\*{3,})$/.test(trimmed) ||
 		/^```|^~~~/.test(trimmed) ||
 		/^\[[^\]]+\]:\s*\S+/.test(trimmed)
@@ -31,18 +37,24 @@ function meaningfulText(line: string): string {
 		return "";
 	}
 
-	return trimmed
-		.replace(
-			/^(?:#{1,6}\s+|(?:>\s*)+|[-*+]\s+(?:\[[ xX]\]\s+)?|\d+[.)]\s+)/,
-			"",
-		)
-		.replace(/!?\[[^\]]*\]\([^)]*\)/g, "")
-		.replace(/!?\[[^\]]*\]\[[^\]]*\]/g, "")
-		.replace(/!?\[\[[^\]]*\]\]/g, "")
-		.replace(/<https?:\/\/[^>]+>/g, "")
-		.replace(/https?:\/\/\S+/g, "")
-		.replace(/<[^>]+>/g, "")
-		.replace(/[*_~`]/g, "")
-		.replace(/\s+/g, " ")
-		.trim();
+	return (
+		trimmed
+			.replace(
+				// Drop heading, quote, list, ordered-list, and task-list prefixes.
+				/^(?:#{1,6}\s+|(?:>\s*)+|[-*+]\s+(?:\[[ xX]\]\s+)?|\d+[.)]\s+)/,
+				"",
+			)
+			// Links, images, and embeds contribute no visible title text here.
+			.replace(/!?\[[^\]]*\]\([^)]*\)/g, "")
+			.replace(/!?\[[^\]]*\]\[[^\]]*\]/g, "")
+			.replace(/!?\[\[[^\]]*\]\]/g, "")
+			// Strip autolinks, bare URLs, and HTML tags.
+			.replace(/<https?:\/\/[^>]+>/g, "")
+			.replace(/https?:\/\/\S+/g, "")
+			.replace(/<[^>]+>/g, "")
+			// Remove inline emphasis/code markers, then fold whitespace.
+			.replace(/[*_~`]/g, "")
+			.replace(/\s+/g, " ")
+			.trim()
+	);
 }
